requirements.txt
snowflake-connector-python>=3.10.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.0
scipy>=1.13.1
matplotlib>=3.9.0
seaborn>=0.13.2
vaderSentiment>=3.3.2
python-dotenv>=1.0.1

